pyproject.toml
[tool.poetry]
name = "online-mart-api"
version = "0.1.0"
description = "Online Mart API"
authors = ["HR-Suriya <130052335+HR-Suriya@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.6"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlmodel = "^0.0.22"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
